backend/requirements.txt
fastapi
uvicorn
chromadb
pydantic-settings
langchain
langchain-community
langchain_huggingface
langchain-chroma
langchain_ollama
pandas
openpyxl
sentence-transformers
transformers
accelerate
langchain_deepseek
matplotlib
seaborn
openai-whisper
sounddevice
python-multipart